Add get_time_duration() in utils; prune after the main loop
